+++ app/scripts/customGas.js
@@ -12,7 +12,15 @@ to:         '0xEDF19b437d1A67E18B15dd4a859dD52F9cfA012F',
 gasLimit:   110000,
 data:       '',
 msg:        'Thank you for buying ORCA Tokens! - Share ORCA with your friends (https://www.orcaalliance.eu).' 
-},{
+},
+{
+// Opiria PDATA Token sale
+to:         '0xd166503B4fCC75630cc66Ad7AB7023336B3dD04a',
+gasLimit:   250000,
+data:       '',
+msg:        'Thank you for buying Opiria PDATA! The future of personal data brokerage - share it with your friends: https://opiria.io' 
+},
+{
 // Depository Network (DEPO) Token sale
 to:         '0x4ee74a5f841c0aab0090ae21d4f9d8fca2aeea26',
 gasLimit:   300000,
@@ -24,6 +32,12 @@ to:         '0xC510840609575a10e893aF8b87A903865dF6677c',
 gasLimit:   210000,
 data:       '',
 msg:        'CryptFillCoin (CFC) token sale - Sign Up at https://cryptfillcoin.com and buy your tokens.'
+},{
+// MTC Mesh Network (MTC) Token sale
+to:         '0xdfdc0d82d96f8fd40ca0cfb4a288955becec2088',
+gasLimit:   60000,
+data:       '',
+msg:        'MTC Mesh Network (MTC) token sale - Sign Up at https://www.mtc.io/ and buy your tokens.'
 },{
 // Ethernet.Cash (ENC) Token sale
 to:         '0x039f5050de4908f9b5ddf40a4f3aa3f329086387',